Add files via upload
Title:Model-Based Systems Engineering Supporting Integrated Modeling and Optimization of Radar Cabin Layout
Abstract:This study proposes a semantic modeling approach based on the KARMA language, which is used to build a system model for the layout design of a UAV radar cabin 
based on Model-Based Systems Engineering (MBSE). Furthermore, the approach is transformed into a Genetic Algorithm (GA) in MATLAB, enabling automatic optimization based on
the generated layout information. The results demonstrate the effectiveness and applicability of the proposed method.